refactory of the code as flask application package
