+++ applications/rag/frontend/container/Dockerfile
@@ -14,32 +14,14 @@
 
 FROM python:3.12.2
 
-ENV DEBIAN_FRONTEND=noninteractive 
-
-RUN curl https://dl.google.com/dl/cloudsdk/release/google-cloud-sdk.tar.gz > /tmp/google-cloud-sdk.tar.gz
-
-RUN mkdir -p /usr/local/gcloud \
-    && tar -C /usr/local/gcloud -xvf /tmp/google-cloud-sdk.tar.gz \
-    && /usr/local/gcloud/google-cloud-sdk/install.sh
-
-ENV PATH $PATH:/usr/local/gcloud/google-cloud-sdk/bin
-
 ADD ./ /workspace/frontend
 WORKDIR /workspace/frontend
 
 RUN pip install -r requirements.txt
 
-
-COPY ./service-account.json /app/
-
-ENV GOOGLE_APPLICATION_CREDENTIALS=/app/service-account.json
-
-RUN gcloud auth activate-service-account --key-file=/app/service-account.json 
-
-ENV PORT=8080
-EXPOSE $PORT
+EXPOSE 8080
 
 ENV FLASK_APP=/workspace/frontend/main.py
-
+ENV PYTHONPATH=.
 # Run the application with Gunicorn
-CMD ["gunicorn", "-w", "4", "-b", "0.0.0.0:$PORT", "main:app"]
+CMD ["gunicorn", "-w", "4", "-b", "0.0.0.0:8080", "main:app"]
