+++ applications/rag/frontend/container/application/rag_langchain/rag_chain.py
